More code review feedback (moved Generator to src/main)
