+++ pom.xml
@@ -187,24 +187,6 @@
         </executions>
       </plugin>
       <plugin>
-        <artifactId>maven-compiler-plugin</artifactId>
-        <version>3.1</version>
-        <executions>
-          <execution>
-            <id>GeneratorCompile</id>
-            <phase>generate-test-sources</phase>
-            <goals>
-              <goal>testCompile</goal>
-            </goals>
-            <configuration>
-              <testIncludes>
-                <testInclude>**/Generator.java</testInclude>
-              </testIncludes>
-            </configuration>
-          </execution>
-        </executions>
-      </plugin>
-      <plugin>
         <groupId>org.codehaus.mojo</groupId>
         <artifactId>exec-maven-plugin</artifactId>
         <version>1.6.0</version>
@@ -217,11 +199,7 @@
           </execution>
         </executions>
         <configuration>
-          <mainClass>org.reaktivity.nukleus.maven.plugin.internal.generator.Generator</mainClass>
-          <classpathScope>test</classpathScope>
-          <additionalClasspathElements>
-            <additionalClasspathElement>${project.build.directory}/test-classes</additionalClasspathElement>
-          </additionalClasspathElements>
+          <mainClass>org.reaktivity.nukleus.maven.plugin.internal.Generator</mainClass>
           <testSourceRoot>${project.build.directory}/generated-test-sources/test-reaktivity</testSourceRoot>
         </configuration>
       </plugin>
