+++ src/main/java/org/reaktivity/nukleus/maven/plugin/internal/AbstractMojo.java
@@ -15,38 +15,18 @@
  */
 package org.reaktivity.nukleus.maven.plugin.internal;
 
-import static java.util.stream.Collectors.toSet;
-
 import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
 import java.net.URI;
 import java.net.URL;
 import java.net.URLClassLoader;
-import java.util.HashSet;
-import java.util.LinkedHashSet;
 import java.util.LinkedList;
 import java.util.List;
-import java.util.Set;
-import java.util.SortedSet;
-import java.util.TreeSet;
 
-import org.antlr.v4.runtime.ANTLRInputStream;
-import org.antlr.v4.runtime.BailErrorStrategy;
-import org.antlr.v4.runtime.CommonTokenStream;
-import org.antlr.v4.runtime.RecognitionException;
-import org.antlr.v4.runtime.Token;
-import org.antlr.v4.runtime.misc.ParseCancellationException;
 import org.apache.maven.artifact.DependencyResolutionRequiredException;
 import org.apache.maven.plugins.annotations.Parameter;
 import org.apache.maven.project.MavenProject;
-import org.reaktivity.nukleus.maven.plugin.internal.ast.AstNode;
 import org.reaktivity.nukleus.maven.plugin.internal.ast.AstSpecificationNode;
-import org.reaktivity.nukleus.maven.plugin.internal.ast.AstStructNode;
-import org.reaktivity.nukleus.maven.plugin.internal.ast.parse.AstParser;
-import org.reaktivity.nukleus.maven.plugin.internal.parser.NukleusLexer;
-import org.reaktivity.nukleus.maven.plugin.internal.parser.NukleusParser;
-import org.reaktivity.nukleus.maven.plugin.internal.parser.NukleusParser.SpecificationContext;
 
 public abstract class AbstractMojo extends org.apache.maven.plugin.AbstractMojo
 {
@@ -62,85 +42,20 @@ public abstract class AbstractMojo extends org.apache.maven.plugin.AbstractMojo
     @Parameter(required = true)
     protected String scopeNames;
 
+    private Parser parser = new Parser()
+            .debug(getLog()::debug)
+            .error(getLog()::error)
+            .warn(getLog()::warn);
+
     protected abstract void executeImpl() throws IOException;
 
     protected final List<AstSpecificationNode> parseAST(
         List<String> targetScopes) throws IOException
     {
-        ClassLoader loader = createLoader();
-
-        List<AstSpecificationNode> specifications = new LinkedList<>();
-        SortedSet<String> parsedResourceNames = new TreeSet<>();
-        Set<String> remainingScopes = new LinkedHashSet<>(targetScopes);
-        while (!remainingScopes.isEmpty())
-        {
-            String remainingScope = remainingScopes.iterator().next();
-            remainingScopes.remove(remainingScope);
-            String resourceName = remainingScope.replaceAll("([^:]+).*", "$1.idl");
-            if (parsedResourceNames.add(resourceName))
-            {
-                getLog().debug("loading: " + resourceName);
-
-                URL resource = loader.getResource(resourceName);
-                if (resource == null)
-                {
-                    getLog().warn(String.format("Resource %s not found", resourceName));
-                    continue;
-                }
-
-                AstSpecificationNode specification = parseSpecification(resourceName, resource);
-                specifications.add(specification);
-
-                Set<String> referencedTypes = specification.accept(new ReferencedTypeResolver());
-                getLog().debug("referenced types: " + referencedTypes);
-
-                String regex = "((:?[^:]+(?:\\:\\:[^:]+)*)?)\\:\\:[^:]+";
-                Set<String> referencedScopes = referencedTypes.stream()
-                                                              .map(t -> t.replaceAll(regex, "$1"))
-                                                              .collect(toSet());
-                getLog().debug("referenced scopes: " + referencedScopes);
-
-                remainingScopes.addAll(referencedScopes);
-            }
-        }
-        return specifications;
+        return parser.parseAST(targetScopes, createLoader());
     }
 
-    private AstSpecificationNode parseSpecification(
-        String resourceName,
-        URL resource) throws IOException
-    {
-        try (InputStream input = resource.openStream())
-        {
-            ANTLRInputStream ais = new ANTLRInputStream(input);
-            NukleusLexer lexer = new NukleusLexer(ais);
-            CommonTokenStream tokens = new CommonTokenStream(lexer);
-            NukleusParser parser = new NukleusParser(tokens);
-            parser.setErrorHandler(new BailErrorStrategy());
-
-            SpecificationContext ctx = parser.specification();
-            return new AstParser().visitSpecification(ctx);
-        }
-        catch (ParseCancellationException ex)
-        {
-            Throwable cause = ex.getCause();
-            if (cause instanceof RecognitionException)
-            {
-                RecognitionException re = (RecognitionException) cause;
-                Token token = re.getOffendingToken();
-                if (token != null)
-                {
-                    String message = String.format("Parse failed in %s at %d:%d on \"%s\"",
-                            resourceName, token.getLine(), token.getCharPositionInLine(), token.getText());
-                    getLog().error(message);
-                }
-            }
-
-            throw ex;
-        }
-    }
-
-    private ClassLoader createLoader() throws IOException
+    ClassLoader createLoader() throws IOException
     {
         List<URL> resourcePath = new LinkedList<>();
 
@@ -151,8 +66,8 @@ public abstract class AbstractMojo extends org.apache.maven.plugin.AbstractMojo
         {
             for (Object resourcePathEntry : project.getTestClasspathElements())
             {
-                File reosourcePathFile = new File(resourcePathEntry.toString());
-                URI resourcePathURI = reosourcePathFile.getAbsoluteFile().toURI();
+                File resourcePathFile = new File(resourcePathEntry.toString());
+                URI resourcePathURI = resourcePathFile.getAbsoluteFile().toURI();
                 resourcePath.add(URI.create(String.format("jar:%s!/META-INF/reaktivity/", resourcePathURI)).toURL());
             }
         }
@@ -165,29 +80,5 @@ public abstract class AbstractMojo extends org.apache.maven.plugin.AbstractMojo
 
         ClassLoader parent = getClass().getClassLoader();
         return new URLClassLoader(resourcePath.toArray(new URL[resourcePath.size()]), parent);
-    }
-
-    private static final class ReferencedTypeResolver extends AstNode.Visitor<Set<String>>
-    {
-        private final Set<String> qualifiedNames = new HashSet<>();
-
-        @Override
-        public Set<String> visitStruct(
-            AstStructNode structNode)
-        {
-            String supertype = structNode.supertype();
-            if (supertype != null)
-            {
-                qualifiedNames.add(supertype);
-            }
-
-            return super.visitStruct(structNode);
-        }
-
-        @Override
-        protected Set<String> defaultResult()
-        {
-            return qualifiedNames;
-        }
     }
 }
